Change options panel names and icons a bit
