+++ Core/src/org/sleuthkit/autopsy/casemodule/Bundle.properties
@@ -221,5 +221,5 @@ Detail\: \n\
 Cannot open a non-Autopsy config file (at {1}).
 XMLCaseManagement.open.msgDlg.notAutCase.title=Error
 AddImageWizardIngestConfigPanel.CANCEL_BUTTON.text=Cancel
-NewCaseVisualPanel1.rbLocalCase.text=Standalone Case
-NewCaseVisualPanel1.rbSharedCase.text=Shared Case
+NewCaseVisualPanel1.rbLocalCase.text=Single-user
+NewCaseVisualPanel1.rbSharedCase.text=Multi-user
