+++ Core/src/org/sleuthkit/autopsy/casemodule/NewCaseVisualPanel1.java
@@ -151,10 +151,9 @@ final class NewCaseVisualPanel1 extends JPanel implements DocumentListener {
         caseDirTextField = new javax.swing.JTextField();
         rbLocalCase = new javax.swing.JRadioButton();
         rbSharedCase = new javax.swing.JRadioButton();
-        
+
         jLabel1.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        org.openide.awt.Mnemonics.setLocalizedText(jLabel1, org.openide.util.NbBundle
-                .getMessage(NewCaseVisualPanel1.class, "NewCaseVisualPanel1.jLabel1.text_1")); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(jLabel1, org.openide.util.NbBundle.getMessage(NewCaseVisualPanel1.class, "NewCaseVisualPanel1.jLabel1.text_1")); // NOI18N
 
         org.openide.awt.Mnemonics.setLocalizedText(caseNameLabel, org.openide.util.NbBundle.getMessage(NewCaseVisualPanel1.class, "NewCaseVisualPanel1.caseNameLabel.text_1")); // NOI18N
