--- Core/src/org/sleuthkit/autopsy/corecomponents/AutopsyDatabasePanel.java
@@ -1,269 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.sleuthkit.autopsy.corecomponents;
-
-import java.awt.Color;
-import javax.swing.event.DocumentEvent;
-import javax.swing.event.DocumentListener;
-import org.sleuthkit.datamodel.CaseDbConnectionInfo;
-import org.sleuthkit.datamodel.CaseDbConnectionInfo.DbType;
-import org.sleuthkit.autopsy.core.UserPreferences;
-
-public class AutopsyDatabasePanel extends javax.swing.JPanel {
-
-    private AutopsyDatabasePanelController controller;
-    private TextBoxChangedListener textBoxChangedListener;
-
-    /**
-     * Creates new form AutopsyDatabasePanel
-     */
-    public AutopsyDatabasePanel(AutopsyDatabasePanelController theController) {
-        initComponents();
-        controller = theController;
-
-        TextPrompt tpHostnameOrIp = new TextPrompt("Hostname or IP Address", tbHostnameOrIp);
-        TextPrompt tpPortNumber = new TextPrompt("Port Number", tbPortNumber);
-        TextPrompt tpUsername = new TextPrompt("User Name", tbUsername);
-        TextPrompt tpPassword = new TextPrompt("Password", tbPassword);
-
-        tpHostnameOrIp.setForeground(Color.LIGHT_GRAY);
-        tpPortNumber.setForeground(Color.LIGHT_GRAY);
-        tpUsername.setForeground(Color.LIGHT_GRAY);
-        tpPassword.setForeground(Color.LIGHT_GRAY);
-
-        float alpha = 0.9f; // Mostly opaque
-        tpHostnameOrIp.changeAlpha(alpha);
-        tpPortNumber.changeAlpha(alpha);
-        tpUsername.changeAlpha(alpha);
-        tpPassword.changeAlpha(alpha);
-
-        setNetworkDbEnabled(cbExternalDbEnabled.isSelected());
-
-        /// Register for notifications when the text boxes get updated
-        textBoxChangedListener = new TextBoxChangedListener();
-        tbHostnameOrIp.getDocument().addDocumentListener(textBoxChangedListener);
-        tbPortNumber.getDocument().addDocumentListener(textBoxChangedListener);
-        tbUsername.getDocument().addDocumentListener(textBoxChangedListener);
-        tbPassword.getDocument().addDocumentListener(textBoxChangedListener);
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel1 = new javax.swing.JPanel();
-        tbHostnameOrIp = new javax.swing.JTextField();
-        tbPortNumber = new javax.swing.JTextField();
-        tbUsername = new javax.swing.JTextField();
-        tbPassword = new javax.swing.JPasswordField();
-        cbExternalDbEnabled = new javax.swing.JCheckBox();
-        lbOops = new javax.swing.JLabel();
-
-        jPanel1.setBorder(javax.swing.BorderFactory.createEtchedBorder());
-
-        tbHostnameOrIp.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        tbHostnameOrIp.setText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbHostnameOrIp.text")); // NOI18N
-        tbHostnameOrIp.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbHostnameOrIp.toolTipText")); // NOI18N
-
-        tbPortNumber.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        tbPortNumber.setText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbPortNumber.text")); // NOI18N
-        tbPortNumber.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbPortNumber.toolTipText")); // NOI18N
-
-        tbUsername.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        tbUsername.setText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbUsername.text")); // NOI18N
-        tbUsername.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbUsername.toolTipText")); // NOI18N
-
-        tbPassword.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        tbPassword.setText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbPassword.text")); // NOI18N
-        tbPassword.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.tbPassword.toolTipText")); // NOI18N
-
-        org.openide.awt.Mnemonics.setLocalizedText(cbExternalDbEnabled, org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.cbExternalDbEnabled.text")); // NOI18N
-        cbExternalDbEnabled.addItemListener(new java.awt.event.ItemListener() {
-            public void itemStateChanged(java.awt.event.ItemEvent evt) {
-                cbExternalDbEnabledItemStateChanged(evt);
-            }
-        });
-
-        lbOops.setFont(new java.awt.Font("Tahoma", 1, 12)); // NOI18N
-        lbOops.setForeground(new java.awt.Color(255, 0, 0));
-        org.openide.awt.Mnemonics.setLocalizedText(lbOops, org.openide.util.NbBundle.getMessage(AutopsyDatabasePanel.class, "AutopsyDatabasePanel.lbOops.text")); // NOI18N
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addComponent(cbExternalDbEnabled)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 50, Short.MAX_VALUE)
-                        .addComponent(lbOops, javax.swing.GroupLayout.PREFERRED_SIZE, 232, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(tbHostnameOrIp)
-                    .addComponent(tbPortNumber)
-                    .addComponent(tbUsername)
-                    .addComponent(tbPassword))
-                .addContainerGap())
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addGap(6, 6, 6)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(cbExternalDbEnabled)
-                    .addComponent(lbOops, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(tbHostnameOrIp, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(tbPortNumber, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(tbUsername, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(tbPassword, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-        this.setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(67, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(133, Short.MAX_VALUE))
-        );
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * Enables/disables the network database settings, based upon input provided
-     *
-     * @param enabled true means enable, false means disable
-     */
-    private void setNetworkDbEnabled(boolean enabled) {
-        tbHostnameOrIp.setEnabled(enabled);
-        tbPortNumber.setEnabled(enabled);
-        tbUsername.setEnabled(enabled);
-        tbPassword.setEnabled(enabled);
-    }
-
-    private void cbExternalDbEnabledItemStateChanged(java.awt.event.ItemEvent evt) {//GEN-FIRST:event_cbExternalDbEnabledItemStateChanged
-        setNetworkDbEnabled(cbExternalDbEnabled.isSelected());
-        controller.changed();
-    }//GEN-LAST:event_cbExternalDbEnabledItemStateChanged
-
-    void load() {
-        CaseDbConnectionInfo info = UserPreferences.getDatabaseConnectionInfo();
-        tbHostnameOrIp.setText(info.getHost());
-        tbPortNumber.setText(info.getPort());
-        tbUsername.setText(info.getUserName());
-        tbPassword.setText(info.getPassword());
-        if (info.getDbType() == DbType.UNKNOWN) {
-            cbExternalDbEnabled.setSelected(false);
-        } else {
-            cbExternalDbEnabled.setSelected(true);
-        }
-
-    }
-
-    void store() {
-
-        DbType dbType = DbType.UNKNOWN;
-
-        if (cbExternalDbEnabled.isSelected()) {
-            dbType = DbType.POSTGRESQL;
-        }
-
-        CaseDbConnectionInfo info = new CaseDbConnectionInfo(
-                tbHostnameOrIp.getText(),
-                tbPortNumber.getText(),
-                tbUsername.getText(),
-                new String(tbPassword.getPassword()),
-                dbType);
-
-        UserPreferences.setDatabaseConnectionInfo(info);
-    }
-
-    /**
-     * Validates that the form is filled out correctly for our usage.
-     *
-     * @return true if it's okay, false otherwise.
-     */
-    boolean valid() {
-        boolean result = false;
-        String text = "";
-        if (cbExternalDbEnabled.isSelected()) {
-            try {
-                if (tbHostnameOrIp.getText().isEmpty()
-                        || tbPortNumber.getText().isEmpty()
-                        || tbUsername.getText().isEmpty()
-                        || tbPassword.getPassword().length == 0) {
-                    // We don't even have everything filled out
-                    result = false;
-                    text = "Fill in all values";
-                } else {
-                    int value = Integer.parseInt(tbPortNumber.getText());
-                    if (value < 0 || value > 65535) { // valid port numbers
-                        result = false; /// port number is invalid
-                        text = "Invalid port number";
-                    } else {
-                        result = true;
-                    }
-                }
-            } catch (Exception ex) {
-                result = false;
-                text = "Invalid port number";
-            }
-        } else {
-            result = true;
-        }
-        lbOops.setText(text);
-        return result;
-    }
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JCheckBox cbExternalDbEnabled;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JLabel lbOops;
-    private javax.swing.JTextField tbHostnameOrIp;
-    private javax.swing.JPasswordField tbPassword;
-    private javax.swing.JTextField tbPortNumber;
-    private javax.swing.JTextField tbUsername;
-    // End of variables declaration//GEN-END:variables
-
-    /**
-     * Used to listen for changes in text boxes. It lets the panel know things
-     * have been updated and that validation needs to happen.
-     */
-    class TextBoxChangedListener implements DocumentListener {
-
-        @Override
-        public void changedUpdate(DocumentEvent e) {
-            controller.changed();
-        }
-
-        @Override
-        public void insertUpdate(DocumentEvent e) {
-            controller.changed();
-        }
-
-        @Override
-
-        public void removeUpdate(DocumentEvent e) {
-            controller.changed();
-        }
-    }
-}
