+++ Core/src/org/sleuthkit/autopsy/corecomponents/AutopsyMultiUserSettingsPanel.java
@@ -1,0 +1,357 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package org.sleuthkit.autopsy.corecomponents;
+
+import java.awt.Color;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
+import org.sleuthkit.datamodel.CaseDbConnectionInfo;
+import org.sleuthkit.datamodel.CaseDbConnectionInfo.DbType;
+import org.sleuthkit.autopsy.core.UserPreferences;
+
+public class AutopsyMultiUserSettingsPanel extends javax.swing.JPanel {
+
+    private AutopsyMultiUserSettingsPanelController controller;
+    private TextBoxChangedListener textBoxChangedListener;
+
+    /**
+     * Creates new form AutopsyMultiUserSettingsPanel
+     */
+    public AutopsyMultiUserSettingsPanel(AutopsyMultiUserSettingsPanelController theController) {
+        initComponents();
+        controller = theController;
+
+        TextPrompt tpHostnameOrIp = new TextPrompt("Hostname or IP Address", tbHostnameOrIp);
+        TextPrompt tpPortNumber = new TextPrompt("Port Number", tbPortNumber);
+        TextPrompt tpUsername = new TextPrompt("User Name", tbUsername);
+        TextPrompt tpPassword = new TextPrompt("Password", tbPassword);
+
+        tpHostnameOrIp.setForeground(Color.LIGHT_GRAY);
+        tpPortNumber.setForeground(Color.LIGHT_GRAY);
+        tpUsername.setForeground(Color.LIGHT_GRAY);
+        tpPassword.setForeground(Color.LIGHT_GRAY);
+
+        float alpha = 0.9f; // Mostly opaque
+        tpHostnameOrIp.changeAlpha(alpha);
+        tpPortNumber.changeAlpha(alpha);
+        tpUsername.changeAlpha(alpha);
+        tpPassword.changeAlpha(alpha);
+
+        setNetworkDbEnabled(cbExternalDbEnabled.isSelected());
+
+        /// Register for notifications when the text boxes get updated
+        textBoxChangedListener = new TextBoxChangedListener();
+        tbHostnameOrIp.getDocument().addDocumentListener(textBoxChangedListener);
+        tbPortNumber.getDocument().addDocumentListener(textBoxChangedListener);
+        tbUsername.getDocument().addDocumentListener(textBoxChangedListener);
+        tbPassword.getDocument().addDocumentListener(textBoxChangedListener);
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        pnOverallPanel = new javax.swing.JPanel();
+        pnExternalDatabase = new javax.swing.JPanel();
+        tbHostnameOrIp = new javax.swing.JTextField();
+        tbPortNumber = new javax.swing.JTextField();
+        tbUsername = new javax.swing.JTextField();
+        tbPassword = new javax.swing.JPasswordField();
+        lbOops = new javax.swing.JLabel();
+        lbExternalDatabase = new javax.swing.JLabel();
+        pnSolrSettings = new javax.swing.JPanel();
+        lbSolrSettings = new javax.swing.JLabel();
+        pnMessagingSettings = new javax.swing.JPanel();
+        lbMessagingSettings = new javax.swing.JLabel();
+        cbExternalDbEnabled = new javax.swing.JCheckBox();
+
+        pnExternalDatabase.setBorder(javax.swing.BorderFactory.createEtchedBorder());
+
+        tbHostnameOrIp.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        tbHostnameOrIp.setText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbHostnameOrIp.text")); // NOI18N
+        tbHostnameOrIp.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbHostnameOrIp.toolTipText")); // NOI18N
+
+        tbPortNumber.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        tbPortNumber.setText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbPortNumber.text")); // NOI18N
+        tbPortNumber.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbPortNumber.toolTipText")); // NOI18N
+
+        tbUsername.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        tbUsername.setText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbUsername.text")); // NOI18N
+        tbUsername.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbUsername.toolTipText")); // NOI18N
+
+        tbPassword.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        tbPassword.setText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbPassword.text")); // NOI18N
+        tbPassword.setToolTipText(org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.tbPassword.toolTipText")); // NOI18N
+
+        lbOops.setFont(new java.awt.Font("Tahoma", 1, 12)); // NOI18N
+        lbOops.setForeground(new java.awt.Color(255, 0, 0));
+        org.openide.awt.Mnemonics.setLocalizedText(lbOops, org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.lbOops.text")); // NOI18N
+        lbOops.setVerticalAlignment(javax.swing.SwingConstants.BOTTOM);
+
+        lbExternalDatabase.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(lbExternalDatabase, org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.lbExternalDatabase.text")); // NOI18N
+        lbExternalDatabase.setVerticalAlignment(javax.swing.SwingConstants.TOP);
+
+        javax.swing.GroupLayout pnExternalDatabaseLayout = new javax.swing.GroupLayout(pnExternalDatabase);
+        pnExternalDatabase.setLayout(pnExternalDatabaseLayout);
+        pnExternalDatabaseLayout.setHorizontalGroup(
+            pnExternalDatabaseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnExternalDatabaseLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(pnExternalDatabaseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(tbHostnameOrIp)
+                    .addComponent(tbPortNumber)
+                    .addComponent(tbUsername)
+                    .addComponent(tbPassword)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnExternalDatabaseLayout.createSequentialGroup()
+                        .addGap(0, 1, Short.MAX_VALUE)
+                        .addComponent(lbExternalDatabase)
+                        .addGap(18, 18, 18)
+                        .addComponent(lbOops, javax.swing.GroupLayout.PREFERRED_SIZE, 320, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addContainerGap())
+        );
+        pnExternalDatabaseLayout.setVerticalGroup(
+            pnExternalDatabaseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnExternalDatabaseLayout.createSequentialGroup()
+                .addGroup(pnExternalDatabaseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(pnExternalDatabaseLayout.createSequentialGroup()
+                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(lbExternalDatabase))
+                    .addComponent(lbOops, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(tbHostnameOrIp, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(tbPortNumber, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(tbUsername, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(tbPassword, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap())
+        );
+
+        pnSolrSettings.setBorder(javax.swing.BorderFactory.createEtchedBorder());
+
+        lbSolrSettings.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(lbSolrSettings, org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.lbSolrSettings.text")); // NOI18N
+
+        javax.swing.GroupLayout pnSolrSettingsLayout = new javax.swing.GroupLayout(pnSolrSettings);
+        pnSolrSettings.setLayout(pnSolrSettingsLayout);
+        pnSolrSettingsLayout.setHorizontalGroup(
+            pnSolrSettingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnSolrSettingsLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(lbSolrSettings)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        pnSolrSettingsLayout.setVerticalGroup(
+            pnSolrSettingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnSolrSettingsLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(lbSolrSettings)
+                .addContainerGap(75, Short.MAX_VALUE))
+        );
+
+        pnMessagingSettings.setBorder(javax.swing.BorderFactory.createEtchedBorder());
+
+        lbMessagingSettings.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(lbMessagingSettings, org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.lbMessagingSettings.text")); // NOI18N
+
+        javax.swing.GroupLayout pnMessagingSettingsLayout = new javax.swing.GroupLayout(pnMessagingSettings);
+        pnMessagingSettings.setLayout(pnMessagingSettingsLayout);
+        pnMessagingSettingsLayout.setHorizontalGroup(
+            pnMessagingSettingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnMessagingSettingsLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(lbMessagingSettings)
+                .addContainerGap(392, Short.MAX_VALUE))
+        );
+        pnMessagingSettingsLayout.setVerticalGroup(
+            pnMessagingSettingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnMessagingSettingsLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(lbMessagingSettings)
+                .addContainerGap(74, Short.MAX_VALUE))
+        );
+
+        org.openide.awt.Mnemonics.setLocalizedText(cbExternalDbEnabled, org.openide.util.NbBundle.getMessage(AutopsyMultiUserSettingsPanel.class, "AutopsyMultiUserSettingsPanel.cbExternalDbEnabled.text")); // NOI18N
+        cbExternalDbEnabled.addItemListener(new java.awt.event.ItemListener() {
+            public void itemStateChanged(java.awt.event.ItemEvent evt) {
+                cbExternalDbEnabledItemStateChanged(evt);
+            }
+        });
+
+        javax.swing.GroupLayout pnOverallPanelLayout = new javax.swing.GroupLayout(pnOverallPanel);
+        pnOverallPanel.setLayout(pnOverallPanelLayout);
+        pnOverallPanelLayout.setHorizontalGroup(
+            pnOverallPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(pnOverallPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(pnOverallPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(cbExternalDbEnabled)
+                    .addGroup(pnOverallPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                        .addComponent(pnSolrSettings, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(pnMessagingSettings, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addComponent(pnExternalDatabase, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addContainerGap(14, Short.MAX_VALUE))
+        );
+        pnOverallPanelLayout.setVerticalGroup(
+            pnOverallPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, pnOverallPanelLayout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(cbExternalDbEnabled)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(pnExternalDatabase, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(pnSolrSettings, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(pnMessagingSettings, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(143, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGap(1, 1, 1)
+                .addComponent(pnOverallPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(pnOverallPanel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    /**
+     * Enables/disables the network database settings, based upon input provided
+     *
+     * @param enabled true means enable, false means disable
+     */
+    private void setNetworkDbEnabled(boolean enabled) {
+        tbHostnameOrIp.setEnabled(enabled);
+        tbPortNumber.setEnabled(enabled);
+        tbUsername.setEnabled(enabled);
+        tbPassword.setEnabled(enabled);
+    }
+
+    private void cbExternalDbEnabledItemStateChanged(java.awt.event.ItemEvent evt) {//GEN-FIRST:event_cbExternalDbEnabledItemStateChanged
+        setNetworkDbEnabled(cbExternalDbEnabled.isSelected());
+        controller.changed();
+    }//GEN-LAST:event_cbExternalDbEnabledItemStateChanged
+
+    void load() {
+        CaseDbConnectionInfo info = UserPreferences.getDatabaseConnectionInfo();
+        tbHostnameOrIp.setText(info.getHost());
+        tbPortNumber.setText(info.getPort());
+        tbUsername.setText(info.getUserName());
+        tbPassword.setText(info.getPassword());
+        if (info.getDbType() == DbType.UNKNOWN) {
+            cbExternalDbEnabled.setSelected(false);
+        } else {
+            cbExternalDbEnabled.setSelected(true);
+        }
+
+    }
+
+    void store() {
+
+        DbType dbType = DbType.UNKNOWN;
+
+        if (cbExternalDbEnabled.isSelected()) {
+            dbType = DbType.POSTGRESQL;
+        }
+
+        CaseDbConnectionInfo info = new CaseDbConnectionInfo(
+                tbHostnameOrIp.getText(),
+                tbPortNumber.getText(),
+                tbUsername.getText(),
+                new String(tbPassword.getPassword()),
+                dbType);
+
+        UserPreferences.setDatabaseConnectionInfo(info);
+    }
+
+    /**
+     * Validates that the form is filled out correctly for our usage.
+     *
+     * @return true if it's okay, false otherwise.
+     */
+    boolean valid() {
+        boolean result = false;
+        String text = "";
+        if (cbExternalDbEnabled.isSelected()) {
+            try {
+                if (tbHostnameOrIp.getText().isEmpty()
+                        || tbPortNumber.getText().isEmpty()
+                        || tbUsername.getText().isEmpty()
+                        || tbPassword.getPassword().length == 0) {
+                    // We don't even have everything filled out
+                    result = false;
+                    text = "Fill in all values";
+                } else {
+                    int value = Integer.parseInt(tbPortNumber.getText());
+                    if (value < 0 || value > 65535) { // valid port numbers
+                        result = false; /// port number is invalid
+                        text = "Invalid port number";
+                    } else {
+                        result = true;
+                    }
+                }
+            } catch (Exception ex) {
+                result = false;
+                text = "Invalid port number";
+            }
+        } else {
+            result = true;
+        }
+        lbOops.setText(text);
+        return result;
+    }
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JCheckBox cbExternalDbEnabled;
+    private javax.swing.JLabel lbExternalDatabase;
+    private javax.swing.JLabel lbMessagingSettings;
+    private javax.swing.JLabel lbOops;
+    private javax.swing.JLabel lbSolrSettings;
+    private javax.swing.JPanel pnExternalDatabase;
+    private javax.swing.JPanel pnMessagingSettings;
+    private javax.swing.JPanel pnOverallPanel;
+    private javax.swing.JPanel pnSolrSettings;
+    private javax.swing.JTextField tbHostnameOrIp;
+    private javax.swing.JPasswordField tbPassword;
+    private javax.swing.JTextField tbPortNumber;
+    private javax.swing.JTextField tbUsername;
+    // End of variables declaration//GEN-END:variables
+
+    /**
+     * Used to listen for changes in text boxes. It lets the panel know things
+     * have been updated and that validation needs to happen.
+     */
+    class TextBoxChangedListener implements DocumentListener {
+
+        @Override
+        public void changedUpdate(DocumentEvent e) {
+            controller.changed();
+        }
+
+        @Override
+        public void insertUpdate(DocumentEvent e) {
+            controller.changed();
+        }
+
+        @Override
+
+        public void removeUpdate(DocumentEvent e) {
+            controller.changed();
+        }
+    }
+}
