+++ Core/src/org/sleuthkit/autopsy/corecomponents/Bundle.properties
@@ -145,15 +145,18 @@ AutopsyOptionsPanel.jLabelNumThreads.text=Number of threads to use for file inge
 FXVideoPanel.progress.bufferingCancelled=media buffering was canceled
 FXVideoPanel.progress.bufferingInterrupted=media buffering was interrupted
 FXVideoPanel.progress.errorWritingVideoToDisk=Error writing video to disk
-OptionsCategory_Name_Database_Settings=External Database
-OptionsCategory_Keywords_Database_Options=Database Options
-AutopsyDatabasePanel.tbPassword.toolTipText=Enter the password here
-AutopsyDatabasePanel.tbPassword.text=
-AutopsyDatabasePanel.tbUsername.toolTipText=User Name
-AutopsyDatabasePanel.tbUsername.text=
-AutopsyDatabasePanel.tbPortNumber.toolTipText=Port Number
-AutopsyDatabasePanel.tbPortNumber.text=
-AutopsyDatabasePanel.tbHostnameOrIp.toolTipText=Hostname or IP Address
-AutopsyDatabasePanel.tbHostnameOrIp.text=
-AutopsyDatabasePanel.cbExternalDbEnabled.text=Use External Database
-AutopsyDatabasePanel.lbOops.text=
+OptionsCategory_Name_Multi_User_Settings=Multi-user
+OptionsCategory_Keywords_Multi_User_Options=Multi-user Options
+AutopsyMultiUserSettingsPanel.lbExternalDatabase.text=External Database Settings
+AutopsyMultiUserSettingsPanel.lbOops.text=
+AutopsyMultiUserSettingsPanel.cbExternalDbEnabled.text=Enable Multi-user
+AutopsyMultiUserSettingsPanel.tbPassword.toolTipText=Enter the password here
+AutopsyMultiUserSettingsPanel.tbPassword.text=
+AutopsyMultiUserSettingsPanel.tbUsername.toolTipText=User Name
+AutopsyMultiUserSettingsPanel.tbUsername.text=
+AutopsyMultiUserSettingsPanel.tbPortNumber.toolTipText=Port Number
+AutopsyMultiUserSettingsPanel.tbPortNumber.text=
+AutopsyMultiUserSettingsPanel.tbHostnameOrIp.toolTipText=Hostname or IP Address
+AutopsyMultiUserSettingsPanel.tbHostnameOrIp.text=
+AutopsyMultiUserSettingsPanel.lbMessagingSettings.text=Messaging Settings
+AutopsyMultiUserSettingsPanel.lbSolrSettings.text=Solr Settings
